+++ src/variants/evidence/mod.rs
@@ -1,3 +1,9 @@
+// Copyright 2020 Johannes Köster.
+// Licensed under the GNU GPLv3 license (https://opensource.org/licenses/GPL-3.0)
+// This file may not be copied, modified, or distributed
+// except according to those terms.
+
 pub mod bases;
 pub mod insert_size;
+pub mod observation;
 pub mod realignment;
