+++ src/variants/evidence/realignment/edit_distance.rs
@@ -1,3 +1,8 @@
+// Copyright 2020 Johannes Köster.
+// Licensed under the GNU GPLv3 license (https://opensource.org/licenses/GPL-3.0)
+// This file may not be copied, modified, or distributed
+// except according to those terms.
+
 use std::cmp;
 use std::fmt::Debug;
