+++ src/variants/sampling_bias/reads.rs
@@ -1,10 +1,15 @@
+// Copyright 2020 Johannes Köster.
+// Licensed under the GNU GPLv3 license (https://opensource.org/licenses/GPL-3.0)
+// This file may not be copied, modified, or distributed
+// except according to those terms.
+
 use std::cmp;
 
 use bio::stats::LogProb;
 
 use crate::estimation::alignment_properties::AlignmentProperties;
 use crate::variants::sampling_bias::SamplingBias;
-use crate::variants::Variant;
+use crate::variants::types::Variant;
 
 pub trait ReadSamplingBias: Variant + SamplingBias {
     /// Probability to sample read from alt allele given the average feasible positions observed
